Added input validation and forward processing

+++ Logic/Checkers.cs
@@ -35,7 +35,7 @@ namespace Logic
 
         public void MoveChecker(string i_COLrow)   // Assuming that validation was checked by GameRules and only next position was given.
         {
-            int col = -1 , row = -1;
+            short col = -1 , row = -1;
             Engine.TranslateCharPositionToIndices(i_COLrow, ref col, ref  row);
             m_Row = (row != -1) ? Convert.ToInt16(row) : m_Row;
             m_Col = (col != -1) ? Convert.ToInt16(col) : m_Col;
